requirements.txt
streamlit==1.29.0
google-generativeai==0.3.1
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.350
